Repository angepasst

+++ src/main/java/ch/zli/m223/punchclock/service/EntryService.java
@@ -22,7 +22,7 @@ public class EntryService {
         return entryRepository.findAll();
     }
 
-    public void deleteEntry(long id){
+    public void deleteEntry(int id){
         entryRepository.deleteById(id);
     }
